:bug: 캠페인 실행중인 쿠폰 삭제 에러처리 및 일시정지 캠페인 종료일 초과 시 자동 종료 처리

+++ src/main/java/com/web/ecommerce/domain/coupon/service/CouponServiceImpl.java
@@ -1,6 +1,7 @@
 package com.web.ecommerce.domain.coupon.service;
 
 import com.web.ecommerce.domain.campaign.entity.Campaign;
+import com.web.ecommerce.domain.campaign.enums.Status;
 import com.web.ecommerce.domain.campaign.repository.CampaignRepository;
 import com.web.ecommerce.domain.coupon.dto.request.CreateCouponRequest;
 import com.web.ecommerce.domain.coupon.dto.request.UpdateCouponRequest;
@@ -159,6 +160,9 @@ public class CouponServiceImpl implements CouponService {
   public void deleteCoupon(Long couponId) {
     Coupon coupon = couponRepository.findById(couponId)
         .orElseThrow(() -> new CustomException(CouponErrorCode.COUPON_NOT_FOUND));
+    if (campaignRepository.existsByCoupon_IdAndStatusIn(couponId, List.of(Status.IN_PROGRESS, Status.PENDING))) {
+      throw new CustomException(CouponErrorCode.COUPON_IN_USE_BY_CAMPAIGN);
+    }
     couponRepository.delete(coupon);
   }
